+++ src/noshow/database/models.py
@@ -87,6 +87,7 @@ class ApiCallResponse(Base):
     __table_args__ = {"schema": "noshow"}
 
     id: Mapped[int] = mapped_column(Integer, primary_key=True, index=True, init=False)
+    timestamp: Mapped[datetime] = mapped_column(DateTime, nullable=True, init=False)
     call_status: Mapped[str]
     call_outcome: Mapped[str]
     remarks: Mapped[str]