+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - Added Alembic for database migrations
+- Added a timestamp for call response, so we can track when patients were called
 
 
 ## [1.4.5] - 2024-10-21